+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The state of the board is now stored in `BishopState`, instead of a single list in `Game` being modified. This improves code readability and also results in small performance improvements in most cases.
 - More descriptive game results. Use `Game.result` to see the exact way the game ended (null if it's still ongoing). Old getters like `Game.checkmate` still exist but `result` is preferred.
 - `Variant.hands` boolean option replaced with `HandOptions`, allowing for variants where hands are enabled but captured pieces aren't added to them (pieces can now be added through actions - see `Variant.spawn()` example).
+- `PieceType` and `MoveDefinition` are now immutable, and are normalised with `copyWith` methods instead of mutation.
 
 ### 0.6.4
 - Fixed gates being output the wrong way round in FEN strings for fixed gating variants.